backend/requirements.txt
fastapi
uvicorn[standard]
python-chess
requests
httpx
pydantic
python-dotenv
psycopg2-binary
sqlalchemy
